requirements.txt
streamlit
tensorflow-cpu
numpy
Pillow
plotly
opencv-python-headless
h5py
scikit-learn